# openapi.yaml
openapi: 3.0.3
info:
  title: Accounts API
  version: 1.0.0
  description: |
    Customer account service for CRM and partner integrations.
    **Demo baseline** — v1.0.0 (PR demo adds `tier` in v1.1.0, PATCH tier in v1.2.0).

servers:
  - url: http://localhost:3000
    description: Local development

tags:
  - name: Accounts
    description: Customer account records

paths:
  /accounts:
    get:
      operationId: listAccounts
      summary: List all accounts
      tags: [Accounts]
      responses:
        '200':
          description: Account list
          content:
            application/json:
              schema:
                type: object
                properties:
                  count:
                    type: integer
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/Account'
    post:
      operationId: createAccount
      summary: Create account
      tags: [Accounts]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AccountInput'
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Account'
        '400':
          description: Validation error
        '409':
          description: Duplicate ID

  /accounts/{id}:
    get:
      operationId: getAccountById
      summary: Get account by ID
      tags: [Accounts]
      parameters:
        - $ref: '#/components/parameters/AccountId'
      responses:
        '200':
          description: Account found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Account'
        '404':
          description: Not found

components:
  parameters:
    AccountId:
      name: id
      in: path
      required: true
      schema:
        type: string
        example: ACC-1001

  schemas:
    Account:
      type: object
      required: [id, name, industry, status]
      properties:
        id:
          type: string
          example: ACC-1001
        name:
          type: string
          example: Northwind Traders
        industry:
          type: string
          example: Retail
        status:
          type: string
          enum: [active, inactive]

    AccountInput:
      type: object
      required: [id, name, industry]
      properties:
        id:
          type: string
        name:
          type: string
        industry:
          type: string
        status:
          type: string
          enum: [active, inactive]
          default: active